backend/pyproject.toml
[project]
name = "hypecast-backend"
version = "0.1.0"
description = "Backend for Hypecast AI commentary"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.132.0",
    "uvicorn[standard]>=0.41.0",
    "pydantic>=2.12.0",
    "getstream>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "vision-agents[gemini,elevenlabs,getstream]>=0.3.8",
    "google-cloud-storage>=2.14.0",
]

[project.optional-dependencies]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::PendingDeprecationWarning"]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.15.2",
]
